TSC copleted implementation in stats card

+++ lib/models/chaining/tsc/tsc_response_model.dart
@@ -87,7 +87,7 @@ class Estimate {
 class StatInterval {
   String min;
   String max;
-  int battleScore;
+  double battleScore;
   String fairFight;
   dynamic lastUpdated;
